requirements.txt
numpy==1.26.4
pillow==10.3.0
imageio==2.34.0
matplotlib==3.8.4
tqdm==4.66.4
opencv-python==4.9.0.80
tqdm==4.66.4 
opencv-python==4.9.0.80
pytest==7.4.0